requirements.txt
ddgs
openai
python-dotenv
